+++ bauh/gems/appimage/__init__.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Optional
 
 from bauh import __app_name__
-from bauh.api.paths import CONFIG_DIR, TEMP_DIR, CACHE_DIR
+from bauh.api.paths import CONFIG_DIR, TEMP_DIR, CACHE_DIR, BINARIES_DIR
 from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -13,7 +13,7 @@ SUGGESTIONS_FILE = f'https://raw.githubusercontent.com/vinifmor/{__app_name__}-f
 CONFIG_FILE = f'{CONFIG_DIR}/appimage.yml'
 APPIMAGE_CONFIG_DIR = f'{CONFIG_DIR}/appimage'
 UPDATES_IGNORED_FILE = f'{APPIMAGE_CONFIG_DIR}/updates_ignored.txt'
-SYMLINKS_DIR = f'{Path.home()}/.local/bin'
+SYMLINKS_DIR = BINARIES_DIR
 URL_COMPRESSED_DATABASES = f'https://raw.githubusercontent.com/vinifmor/{__app_name__}-files/master/appimage/dbs.tar.gz'
 APPIMAGE_CACHE_DIR = f'{CACHE_DIR}/appimage'
 DATABASE_APPS_FILE = f'{APPIMAGE_CACHE_DIR}/apps.db'